Trying to implement blocked processes, got blue screen

+++ src/kernel/keyboard/consread.h
@@ -1,7 +1,7 @@
 #ifndef _CONSREAD_H_
 #define _CONSREAD_H_
 
-int cons_read(char *string, unsigned long length);
+unsigned long cons_read(char *string, unsigned long length);
 void cons_echo(int on);
 
 #endif
